fix(shell): split sidecar auth monitor
