+++ src-tauri/src/lib.rs
@@ -719,7 +719,7 @@ pub fn run() {
                 ));
             }
 
-            let url = format!("http://127.0.0.1:{}/", port);
+            let url = format!("http://127.0.0.1:{}/?covenCaveToken={}", port, auth_token);
             let main_url = url.parse().expect("valid url");
             pty::trust_main_origin(&main_url);
             if let Err(e) = WebviewWindowBuilder::new(
