+++ src/components/security/sidecar-auth-bridge.tsx
@@ -55,51 +55,3 @@ const SIDECAR_AUTH_BRIDGE = `
 export function SidecarAuthBridge() {
   return <script dangerouslySetInnerHTML={{ __html: SIDECAR_AUTH_BRIDGE }} />;
 }
-
-// ---------------------------------------------------------------------------
-// SidecarAuthMonitor — client-side companion that surfaces auth failures via
-// the shell-banners channel. The inline script above runs before React hydrates
-// and silently bails out when no token is present; this component detects the
-// same condition on mount and pushes an error banner so the user can act.
-// ---------------------------------------------------------------------------
-
-"use client";
-
-import { useEffect } from "react";
-import { useShellBanners } from "@/lib/shell-banners";
-
-const TOKEN_PARAM = "covenCaveToken";
-const STORAGE_KEY = "coven-cave:sidecar-auth-token";
-const BANNER_ID = "sidecar-auth-failed";
-
-export function SidecarAuthMonitor() {
-  const { pushBanner, dismissBanner } = useShellBanners();
-
-  useEffect(() => {
-    const params = new URLSearchParams(window.location.search);
-    const token =
-      params.get(TOKEN_PARAM) ?? window.sessionStorage.getItem(STORAGE_KEY);
-
-    if (!token) {
-      console.error(
-        "[SidecarAuthBridge] No sidecar auth token found — local APIs will not be authenticated.",
-      );
-      pushBanner({
-        id: BANNER_ID,
-        severity: "error",
-        title: "Sidecar authentication failed — local APIs may not work.",
-        cta: {
-          label: "Open settings",
-          onClick: () => {
-            window.location.href = "/settings";
-          },
-        },
-      });
-    } else {
-      // Token is present — clear any stale failure banner from a previous load.
-      dismissBanner(BANNER_ID);
-    }
-  }, [pushBanner, dismissBanner]);
-
-  return null;
-}
